requirements.txt
pytest
Appium-Python-Client
selenium
colorlog
pyyaml
